% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions2.R
\name{adjust_gr}
\alias{adjust_gr}
\title{Adjust GRanges}
\usage{
adjust_gr(gr, upstream = 0, downstream = 0)
}
\arguments{
\item{gr}{A GRanges object}

\item{upstream}{An integer specifying the the number of bases to adjust each region upstream of its start. Default is 0 bases}

\item{Downstream}{An An integer specifying the number of bases to adjust each region downstream of its end. Default is 0 bases}
}
\value{
A GRanges object.
}
\description{
Adjust ranges in a GRanges object upstream and downstream by specified numbers of bases. Takes strand into account.
}
