% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{bedtools_intersect}
\alias{bedtools_intersect}
\title{Return the regions in a query GRanges which overlap a subject GRanges, keeping the metadata from query regions}
\usage{
bedtools_intersect(query_gr, subject_gr)
}
\arguments{
\item{gr1}{A GRanges object.}

\item{gr2}{A GRanges object.}
}
\value{
A GRanges object.
}
\description{
Return the regions in a query GRanges which overlap a subject GRanges, keeping the metadata from query regions
}
