% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_bed_regions_in_bigwigs.R
\name{bigwig_summarize_over_regions}
\alias{bigwig_summarize_over_regions}
\title{Summarize values in bigWig files for a set of genomic regions}
\usage{
bigwig_summarize_over_regions(
  bw_filepaths,
  bed_filepath = NULL,
  gr = NULL,
  statistic,
  column_names = NULL,
  ncores = 1
)
}
\arguments{
\item{bw_filepaths}{A vector of filepaths for bigWig files.}

\item{bed_filepath}{Filepath for a BED file with genomic regions of interest.
One of either bed_filepath or grs must be provided.}

\item{gr}{A GRanges object. One of either bed_filepath or gr must be provided.}

\item{statistic}{Statistic to compute. One of "mean" (average in bigWig files over just covered bases in regions),
"mean0" (average over bases with non-covered bases counting as zeroes) or "sum".}

\item{column_names}{A vector of names to use as column names of the output table.
Default is to use the base name of bw_filepaths.}

\item{ncores}{Number of cores to use. Default is 1}
}
\value{
A data.frame with the results.
}
\description{
Calculates the mean (with or without non-covered bases counting as zeros) or the sum of values in bigWig files
for provided regions. Regions can be provided either as a GRanges object or as a path to a BED file.
Columns in the output correspond to input bigWig files and rows correspond to genomic regions.
Uses names of the provided BED file or GRanges as row names so long as they are unique.
Otherwise names regions region_1, region_2, etc.
}
