% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{bin_relative_ranges}
\alias{bin_relative_ranges}
\title{Create relative bins and count the overlaps of relative ranges with relative bins}
\usage{
bin_relative_ranges(relative_ranges, bin_start, bin_end, bin_step, category)
}
\arguments{
\item{relative_ranges}{A GRanges object returned by relative_ranges}

\item{bin_start}{Start of bins}

\item{bin_end}{End of bins}

\item{bin_step}{Size of bins}

\item{category}{A category associated with relative_ranges}
}
\value{
A data.frame with the the number of ranges of each category in each bin
}
\description{
Create relative bins and count the overlaps of relative ranges with relative bins
}
