% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{calculate_regions_intersection_list}
\alias{calculate_regions_intersection_list}
\title{Calculate the number of bases in the intersection of all pairs of GRanges from a list}
\usage{
calculate_regions_intersection_list(grl, ignore.strand = T, as.proportion = F)
}
\arguments{
\item{grl}{A list of GRanges objects}

\item{ignore.strand}{A logical value indicating whether strand should be ignored when calculating intersections. Default is TRUE.}

\item{as.proportion}{A logical value indicating whether the overlap should be returned as a proportion of the GRanges indicated by the columns. Default value is FALSE.}
}
\value{
A matrix with the size of the overlaps between all pairs of GRanges in grl
}
\description{
Calculate the number of bases in the intersection of all pairs of GRanges from a list
}
