% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{calculate_regions_jaccard_list}
\alias{calculate_regions_jaccard_list}
\title{Calculate the Jaccard index of the intersections of all pairs of GRanges from a list}
\usage{
calculate_regions_jaccard_list(grl, ignore.strand = T)
}
\arguments{
\item{grl}{A list of GRanges objects}

\item{ignore.strand}{A logical value indicating whether strand should be ignored when calculating intersections. Default is TRUE.}

\item{as.proportion}{A logical value indicating whether the overlap should be returned as a proprtion of gr1. Default value is FALSE.}
}
\value{
A matrix with the size of the overlaps between all pairs of GRanges in grl
}
\description{
Calculate the Jaccard index of the intersections of all pairs of GRanges from a list
}
