% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{calculate_regions_overlap_list}
\alias{calculate_regions_overlap_list}
\title{Calculate the number of ranges of each GRanges object in a list overlapping the others with a minimum overlap proportion threshold}
\usage{
calculate_regions_overlap_list(grl, ignore.strand = T, overlap_threshold = 0.5)
}
\arguments{
\item{grl}{A list of GRanges objects}

\item{ignore.strand}{A logical value indicating whether strand should be ignored when calculating overlaps. Default is TRUE.}

\item{overlap_threshold}{Minimum overlap proprtion of the query ranges with the subject to consider them overlapping. Default is 0.5.}
}
\description{
Calculate the number of ranges of each GRanges object in a list overlapping the others with a minimum overlap proportion threshold
}
