% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions2.R
\name{check_self_overlaps}
\alias{check_self_overlaps}
\title{Check if there are any overlapping regions within a GRanges object}
\usage{
check_self_overlaps(gr, ignore.strand = TRUE)
}
\arguments{
\item{gr}{A GRanges object.}

\item{ignore.strand}{A logical value indicating whether to ignore strand when checking for overlaps.
Default is TRUE.}
}
\value{
FALSE if there are no overlapping regions in gr and TRUE if there are.
}
\description{
Check if there are any overlapping regions within a GRanges object
}
