% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_between_bedgraphs_and_bigwigs.R
\name{convert_bedgraphs_to_bigwigs}
\alias{convert_bedgraphs_to_bigwigs}
\title{Convert begGraph files to bigWig files}
\usage{
convert_bedgraphs_to_bigwigs(
  bg_files,
  genome_build,
  output_dir = ".",
  ncores = 1
)
}
\arguments{
\item{bg_files}{Paths to a directory of bedGraph files}

\item{genome_build}{Either "hg19" or "hg38" depending on the intended output genome}

\item{output_dir}{Name of the directory where created bigWigs will be saved. Default is current working directory}

\item{ncores}{The number of cores to use. Default 1}
}
\value{
None
}
\description{
Convert begGraph files to bigWig files and save in a specified directory.
bedGraph files can be compressed with gzip. Takes ~ 1 minute per file. If more than one file, each file can be processed in parallel.
}
