% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_between_bigwigs_and_bedgraphs.R
\name{convert_bigwigs_to_bedgraphs}
\alias{convert_bigwigs_to_bedgraphs}
\title{Convert bigWig files to bedGraph files}
\usage{
convert_bigwigs_to_bedgraphs(
  bw_files,
  compress_output_bedgraphs = T,
  output_dir = ".",
  ncores = 1
)
}
\arguments{
\item{bw_files}{Paths to a directory of bigwig files}

\item{compress_output_bedgraphs}{A logical value indicating whether to compress output bedGraph files with gzip. Default is TRUE.}

\item{output_dir}{Name of the directory where created bigWigs will be saved. Default is current working directory}

\item{ncores}{The number of cores to use. Default 1}
}
\value{
None
}
\description{
Convert bigWig files to bedGraph files and save in a specified directory. Can compress bedGraphs using gzip if specified.
Takes ~ 1 minute per file. If more than one file, each file can be processed in parallel.
}
