% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{create_relative_bins}
\alias{create_relative_bins}
\title{Create relative bins}
\usage{
create_relative_bins(bin_start, bin_end, bin_step)
}
\arguments{
\item{bin_start}{Start of bins}

\item{bin_end}{End of bins}

\item{bin_step}{Size of bins}
}
\value{
A GRanges object
}
\description{
Create relative bins
}
