% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions2.R
\name{filter_gr}
\alias{filter_gr}
\title{Filter GRanges}
\usage{
filter_gr(gr, column, values, invert = FALSE)
}
\arguments{
\item{gr}{A GRanges object}

\item{column}{The name of a metadata column of gr to filter on.}

\item{values}{A vector of values of column to filter for.}

\item{invert}{Whether to filter for ranges without these values for the specified column. Default is FALSE.}
}
\value{
A GRanges object.
}
\description{
Filter GRanges
}
