% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_convert.R
\name{gene_id_convert}
\alias{gene_id_convert}
\title{Convert gene identifiers}
\usage{
gene_id_convert(gene_ids, input_id_type, output_id_type)
}
\arguments{
\item{gene_ids}{A character vector with gene IDs to convert}

\item{input_id_type}{Type of input gene ID. Options are "ensembl_id", "hgnc_id", "hgnc_symbol" or "entrez_id".}

\item{output_id_type}{Type of output gene ID. Options are "ensembl_id", "hgnc_id", "hgnc_symbol" or "entrez_id".}
}
\value{
None
}
\description{
Convert between Ensembl IDs, HGNC IDs, HGNC symbols and Entrez IDs
}
