% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_convert.R
\name{gene_select_by_chrom}
\alias{gene_select_by_chrom}
\title{Selects the correct gene using chromosome name when multiple genes are returned by HGNChelper::checkGeneSymbols()}
\usage{
gene_select_by_chrom(genes, chromosomes)
}
\arguments{
\item{chromosomes}{A vector of the names of chromosomes associated with each gene in \code{gene_symbols}}

\item{gene_symbols}{A vector of gene names}
}
\value{
None
}
\description{
Selects the correct gene for a given chromosome when multiple genes are returned by HGNChelper::checkGeneSymbols(),
if one and only one of the returned genes is on the indicated chromosome. Returns NA if more than one or none of the genes are on the indicated chromosome.
}
