\name{GenomicTools-package}
\alias{R/GenomicTools-package}
\docType{package}

\title{
  Loose Collection of Expression and Genotype Analysis Tools, including (e)QTL and MDR.
}

\description{
  GenomicTools is a loose collection of analysis tools for gene expression and genotype data. Currently
  it provides methods for eQTL, QTL and MDR analysis. It also contains an MDR ensemble classifier. This
  package is the continuation of the genomic data analysis parts of the GeneticTools package, that is also
  available on Cran.
}

\details{
  \tabular{ll}{
    Package: \tab GenomicTools\cr
    Type: \tab Package\cr
    Version: \tab 0.2.9.7\cr
    Date: \tab 2020-03-09\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer

  Maintainer: Daniel Fischer <daniel.fischer@luke.fi>
}

\keyword{multivariate}

