\name{genotDataVCF}
\alias{genotDataVCF}
\docType{data}
\title{Simulated Genotype Data}
\description{
  Simulated genotype data
}
\usage{data(genotDataVCF)}

\format{
  A \code{vcf} object as provided from \code{importVCF} 
}

\details{
  Small simulated genotype dataset that can be used for testing the method. 
}

\source{
  Own simulation, code can be given upon request. 
}

\examples{
data(genotDataVCF)
}
\keyword{datasets}
 
