\name{getKEGGInformation}
\alias{getKEGGInformation}

\title{
  Get Information from the KEGG database
}

\description{
  This function gets information from the KEGG database.
}
\usage{
  getKEGGInformation(x)
}

\arguments{
  \item{x}{Information to obtain}
}

\details{
  This function extracts information KEGG database.
}
\value{
  Informations
}

\author{
  Daniel Fischer
}

 
