\name{getKEGGModule}
\alias{getKEGGModule}

\title{
  Get module information from the KEGG database
}

\description{
  This function gets module information from the KEGG database.
}
\usage{
  getKEGGModule(module)
}

\arguments{
  \item{module}{Module name}
}

\details{
  This function extracts module information KEGG database.
}
\value{
  A lsit with module information
}

\author{
  Daniel Fischer
}

 
