\name{getKEGGModuleOverview}
\alias{getKEGGModuleOverview}

\title{
  Get module overview from the KEGG database
}

\description{
  This function gets module overview from the KEGG database.
}
\usage{
  getKEGGModuleOverview()
}

\details{
  This function extracts module overview KEGG database.
}
\value{
  A list with module overview information
}

\author{
  Daniel Fischer
}

 
