\name{getKEGGOrthologOverview}
\alias{getKEGGOrthologOverview}

\title{
  Get ortholog overview from the KEGG database
}

\description{
  This function gets ortholog overview from the KEGG database.
}
\usage{
  getKEGGOrthologOverview()
}

\details{
  This function extracts ortholog overview KEGG database.
}
\value{
  A list with ortholog overview information
}

\author{
  Daniel Fischer
}

 
