% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\docType{data}
\name{hg19_ucsc_seqinfo}
\alias{hg19_ucsc_seqinfo}
\title{A list of colour vectors}
\format{
A Seqinfo object for hg19 for UCSC
}
\usage{
hg19_ucsc_seqinfo
}
\description{
A list of colour vectors
}
\keyword{datasets}
