\name{importSTARLog}
\alias{importSTARLog}

\title{
  Import a log-file from the STAR mapper.
}

\description{
  This function imports a log-file from the STAR mapper.
}
\usage{
  importSTARLog(file)
}

\arguments{
  \item{file}{The log filename.}
}

\details{
 This function reads the log files as they are provided by the STAR mapper.  
}
\value{
 A data frame that contains the log file.
}


\author{
  Daniel Fischer
}

\keyword{
  multivariate
}
 
