\name{importVCF}
\alias{importVCF}

\title{
  Import a vcf file.
}

\description{
  This function imports a vcf file.
}
\usage{
  importVCF(file)
}

\arguments{
  \item{file}{The vcf filename.}
}

\details{
 This function imports a vcf file. 
}
\value{
 A \code{vcf} object.
}


\author{
  Daniel Fischer
}

\keyword{
  multivariate
}
 
