% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{individual_proportion_overlaps}
\alias{individual_proportion_overlaps}
\title{#' Calculate the number of ranges in one GRanges object overlapping another
#'
#' @param gr1 A GRanges object
#' @param gr1 A GRanges object
#' @param ignore.strand A logical value indicating whether strand should be ignored when calculating overlap. Default is TRUE.
#' @param as.proportion A logical value indicating whether the overlap should be returned as a proportion of gr1. Default value is FALSE.
#' @return An numeric value
#' @export
calculate_regions_overlaps = function(gr1, gr2, ignore.strand = T, as.proportion = F){

overlap = length(IRanges::subsetByOverlaps(gr1, gr2, ignore.strand = ignore.strand))

if(as.proportion){
return(overlap/length(gr1))
} else {
return(overlap)
}}
\usage{
individual_proportion_overlaps(query_gr, subject_gr, ignore.strand = T)
}
\arguments{
\item{query_gr}{A GRanges object. The proportion of each region overlapping regions from subject_gr will be calcualted.}

\item{subject_gr}{A GRanges object}

\item{ignore.strand}{A logical value indicating whether to ignore strand when calculating overlaps. Default is TRUE.}
}
\value{
A matrix with the size of the overlaps between all pairs of GRanges in grl
}
\description{
}
#' Calculate the number of ranges of each GRanges object in a list overlapping the others
#'
#' @param grl A list of GRanges objects
#' @param ignore.strand A logical value indicating whether strand should be ignored when calculating overlaps. Default is TRUE.
#' @param as.proportion A logical value indicating whether the overlap should be returned as a proprtion of gr1. Default value is FALSE.
#' @return A matrix with the size of the overlaps between all pairs of GRanges in grl
#' @export
calculate_regions_overlap_list = function(grl, ignore.strand = T, as.proportion = F){

overlap_matrix = matrix(NA, nrow = length(grl), ncol = length(grl), dimnames = list(names(grl), names(grl)))

for(i in names(grl)){
for(j in names(grl)){
overlap_matrix\link{i, j} = genomeTools::calculate_regions_overlaps(grl[\link{i}], grl[\link{j}], ignore.strand = ignore.strand, as.proportion = as.proportion)
}
}

return(overlap_matrix)
}
\details{

}
