% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover_functions.R
\name{liftover_files}
\alias{liftover_files}
\title{Liftover bigWig or bedGraph files from one genome build to another}
\usage{
liftover_files(
  input_filepaths,
  input_filetype,
  output_filetype,
  to_genome_build,
  output_directory = ".",
  compress_output_bedGraphs = T,
  ncores = 1
)
}
\arguments{
\item{input_filepaths}{Paths to either bedGraph files or bigWig files to be lifted-over. All files must be of the same type. bedGraph files may be zipped.}

\item{input_filetype}{Either "bw" for bigWig or "bg" for bedGraph.}

\item{output_filetype}{Either "bw" for bigWig or "bg" for bedGraph.}

\item{to_genome_build}{Either "hg19" or "hg38" depending on the intended output genome}

\item{output_directory}{Directory to save output files in. Default is current working directory. If output_directory doesn't exist, it is created.}

\item{compress_output_bedGraphs}{A logical value indicating whether output bedGraph files should be compressed. Default is TRUE.}

\item{ncores}{The number of cores to use to process files in parallel. Default is 1.}

\item{chain}{A "Chain" object to be used with rtracklayer::liftOver.}
}
\value{
None
}
\description{
Liftover either bigwig or bedGraphs from hg19 to hg38 or from hg38 to hg19 and save as a bigWig or a bedGraph.
bedRemoveOverlap is used to remove overlapping regions resulting from the output files after liftover.
Output files have the basename of the input files with "liftover" and either "hg19" or "hg38" appended, depending on value of to_genome_build
}
