% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover_granges.R
\name{liftover_granges}
\alias{liftover_granges}
\title{Liftover a GRanges object}
\usage{
liftover_granges(
  genomic_regions,
  chain,
  remove_non_mapping = T,
  remove_one_to_many_mapping = T,
  remove_many_to_one_mapping = T,
  target_regions_overlap_filter = NULL,
  flatten_ranges = T
)
}
\arguments{
\item{chain}{A "Chain" object to be used with rtracklayer::liftOver.}

\item{remove_non_mapping}{A logical value indicating whether to remove regions in the source genome.
which could not be mapped to the target genome. Default is TRUE.}

\item{remove_one_to_many_mapping}{A logical value indicating whether to remove regions in the source genome.
which map to multiple regions in the target genome. Default is TRUE.}

\item{remove_many_to_one_mapping}{A logical value indicating whether to remove different regions in the source genome.
which map to the same region in the target genome i.e. remove regions which overlap after the liftover. Default is TRUE.}

\item{target_regions_overlap_filter}{An optional GRanges object used to filter the rowRanges by overlaps after liftover,
for example CpG sites from the target genome. Any regions which do not overlap target_regions_overlap_filter will be removed.}

\item{flatten_ranges}{A logical value indicating whether to convert the lifted over ranges from a
GRangesList to GRanges if all remaining source regions can be uniquely mapped to the target genome.}

\item{genomic_region}{A GRanges object.}
}
\value{
A GRanges object
}
\description{
Liftover a GRanges from one genome build to another using a provided liftover chain file
}
