\name{mdrExample}
\alias{mdrExample}
\docType{data}
\title{Example Data for MDR}
\description{
  This is one example data of the Java MDR program.
}
\usage{data(mdrExample)}

\format{
  A data frame. 
}

\details{
  This is an example dataset to show the use of the mdr function.
}

\source{
  https://sourceforge.net/projects/mdr/
}

\examples{
data(mdrExample)
mdrExample[1:5,1:5]
}
\keyword{datasets}
 
