% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{merge_overlapping_regions}
\alias{merge_overlapping_regions}
\title{Merge overlapping regions in a GRanges object}
\usage{
merge_overlapping_regions(gr)
}
\arguments{
\item{gr}{A GRanges object.}
}
\value{
A GRanges object.
}
\description{
Merge overlapping regions in a GRanges object
}
