% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{merge_touching_regions}
\alias{merge_touching_regions}
\title{Merge touching and overlapping regions in a GRanges object}
\usage{
merge_touching_regions(gr)
}
\arguments{
\item{gr}{A GRanges object.}
}
\value{
A GRanges object.
}
\description{
Merge touching and overlapping regions in a GRanges object
}
