\name{print.mdr}
\docType{methods}
\alias{print.mdr}
\alias{print,mdr-method}

\title{Print an mdr Object}
\description{
  Prints an \code{mdr} object.
}
\usage{
 \method{print}{mdr}(x,...)
}

\arguments{
  \item{x}{Object of class \code{mdr}.}
  \item{...}{Additional parameters.}
}

\details{
  The function prints an \code{mdr} object. 
}

\author{Daniel Fischer}

\examples{
# res <- mdr(X=temp,status=status,fold=3,top=20)
# res
}

\keyword{methods}
\keyword{print}
