% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{query_subject_cpg_overlap_test}
\alias{query_subject_cpg_overlap_test}
\title{Perform a Chi-squared test for the overlap of CpG sites in a set of query regions with a set of subject regions comapred to a set of control regions}
\usage{
query_subject_cpg_overlap_test(
  query_regions,
  subject_regions,
  control_regions,
  background_cpgs,
  alternative = "greater"
)
}
\arguments{
\item{query_regions}{A GRanges object}

\item{subject_regions}{A GRanges object}

\item{control_regions}{A GRanges object}

\item{background_cpgs}{A GRanges object with the location of all genomic CpG sites}

\item{alternative}{The alternative hypothesis. Must be one of "two.sided", "greater" or "less". "greater" is default value.}
}
\value{
The results of the chi-squared test
}
\description{
Perform a Chi-squared test for the overlap of CpG sites in a set of query regions with a set of subject regions comapred to a set of control regions
}
