\name{realSNPs}
\alias{realSNPs}
\docType{data}
\title{Simulated SNP Data}
\description{
  Simulated SNP data.
}
\usage{data(realSNPs)}

\format{
  Data format as it is provided from \code{package::snpStats} and its function \code{read.pedfile}.
}

\details{
  Just simulated data for testing purposes of the functions.
}

\source{
  Own imulation, underlying code can be given upon request.
}

\examples{
data(realSNPs)
}
\keyword{datasets}
 
