% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{relative_ranges}
\alias{relative_ranges}
\title{Find locations of ranges relative to a reference positions}
\usage{
relative_ranges(gr, reference_positions, gr_category = NULL)
}
\arguments{
\item{gr}{A GRanges object}

\item{reference_positions}{A GRanges object. Each range should have width 1. Upstream and downstream refer to reference_positions}

\item{gr_category}{A category that ranges belong to}
}
\value{
A GRanges object
}
\description{
Find locations of ranges relative to a reference positions
}
