\name{revcomp}
\alias{revcomp}

\title{
  Get the reverse complement from teh sequence
}

\description{
  This function provides the reverse complement from the nucleotide sequence
}
\usage{
  revcomp(x, rev=TRUE)
}

\arguments{
  \item{x}{Nucleotide sequence}
  \item{rev}{Logical, shall the sequence reverse also}
}

\details{
  This function provides the reverse complement from the nucleotide sequence.
}
\value{
  A vector with the (reverse) complement of the sequence.
}

\author{
  Daniel Fischer
}

 
