\name{seqlength}
\alias{seqlength}

\title{
  Get the sequence length of a fa-object.
}

\description{
  This function determines the length of a fa-object.
}
\usage{
  seqlength(x)
}

\arguments{
  \item{x}{The fa-object}
}

\details{
This is basically a wrapper around the \code{nchar}-function
}
\value{
 An vector containing the sequence lengths of an fa-object.
}

\author{
  Daniel Fischer
}

\keyword{
  multivariate
}
 
