\name{snpData}
\alias{snpData}
\docType{data}
\title{Example SNP data}
\description{
  This is one example data for eQTL data analysis.
}
\usage{data(snpData)}

\format{
  A data frame. 
}

\details{
  This is an example dataset to show the use of the (e)QTL function.
}

\source{
  Simulated
}

\examples{
data(snpData)
}
\keyword{datasets}
 
