% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id_convert.R
\name{update_gene_symbols}
\alias{update_gene_symbols}
\title{Update gene symbols}
\usage{
update_gene_symbols(gene_symbols, ambiguous_as_na = F, remove_na = F)
}
\arguments{
\item{gene_symbols}{A vector of gene names to update}

\item{ambiguous_as_na}{A logical value indicating whether to replace ambiguous results (i.e. results mapping to more than one gene symbol) with NA. Default is FALSE.}

\item{remove_na}{A logical value indicating whether to remove NA values  and sort the returned results, Default if FALSE.}
}
\value{
None
}
\description{
Update symbols using HGNChelper
}
