\name{Xgene}
\alias{Xgene}
\docType{data}
\title{Simulated Expression Data}
\description{
  Simulated expression data.
}
\usage{data(Xgene)}

\format{
  A matrix with 100 rows representing individuals and 13 genes arranged in the columns.
}

\details{
  Just simulated, numerical expression data for testing purposes.
}

\source{
  Own simulation, code can be given upon request.
}

\examples{
data(Xgene)
}
\keyword{datasets}
 
