\name{genotData}
\alias{genotData}
\docType{data}
\title{Simulated Genotype Data}
\description{
  Simulated genotype data
}
\usage{data(genotData)}

\format{
  A \code{PedMap} object as provided from \code{importPED} 
}

\details{
  Small simulated genotype dataset that can be used for testing the method. 
}

\source{
  Own simulation, code can be given upon request. 
}

\examples{
data(genotData)
}
\keyword{datasets}
 
